move back non tensor related types
