+++ tt_metal/api/tt-metalium/experimental/tensor/tensor_types.hpp
@@ -61,25 +61,8 @@ bool is_floating_point(DataType dtype);
 
 bool is_block_float(DataType dtype);
 
-// Specifies Tensor storage type.
-enum class StorageType {
-    HOST = 0,
-    DEVICE = 1,
-};
-
 tt::DataFormat datatype_to_dataformat_converter(DataType datatype);
 tt::tt_metal::DataType dataformat_to_datatype_converter(tt::DataFormat dataformat);
-
-static constexpr std::size_t MAX_NUM_DIMENSIONS = 8;
-
-using Array1D = std::array<uint32_t, 1>;
-using Array2D = std::array<uint32_t, 2>;
-using Array3D = std::array<uint32_t, 3>;
-using Array4D = std::array<uint32_t, 4>;
-using Array5D = std::array<uint32_t, 5>;
-using Array6D = std::array<uint32_t, 6>;
-using Array7D = std::array<uint32_t, 7>;
-using Array8D = std::array<uint32_t, 8>;
 
 struct NdShardSpec {
     Shape shard_shape;
