+++ ttnn/api/ttnn/tensor/types.hpp
@@ -24,3 +24,24 @@ enum class PyDType {
     BOOL
 };
 }
+
+namespace tt::tt_metal {
+
+// Specifies Tensor storage type.
+enum class StorageType {
+    HOST = 0,
+    DEVICE = 1,
+};
+
+static constexpr std::size_t MAX_NUM_DIMENSIONS = 8;
+
+using Array1D = std::array<uint32_t, 1>;
+using Array2D = std::array<uint32_t, 2>;
+using Array3D = std::array<uint32_t, 3>;
+using Array4D = std::array<uint32_t, 4>;
+using Array5D = std::array<uint32_t, 5>;
+using Array6D = std::array<uint32_t, 6>;
+using Array7D = std::array<uint32_t, 7>;
+using Array8D = std::array<uint32_t, 8>;
+
+}  // namespace tt::tt_metal
